[733] Improved Sidebar logic
